Moves teiid version and data type manager fields up to parent classes
* AbstractNodeVisitor
 * Moves the dataTypeManager field from TeiidSqlNodeVisitor to its parent
   since DdlNodeVisitor also requires it.
 * Moves version testing method to parent

* AbstractSequencerTest
 * Moves teiid version from AbstractTSqlSequencerTest to its parent since
   DdlNodeVisitor tests require it.
